requirements.txt
# Core framework
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0

# Auth/OpenAPI direct dependencies
PyJWT==2.10.1
PyYAML==6.0.3
uritemplate==4.2.0
inflection==0.5.1
jsonschema==4.26.0

# Transitive dependencies pinned for reproducibility
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
attrs==26.1.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# Optional for PostgreSQL production deployment
# psycopg2-binary==2.9.9
